Split DataMaster and Flask server into separate modules
